+++ resources/views/components/user-collections/share-form.blade.php
@@ -1,5 +1,7 @@
 <user-collections-share-form
         action="{{ $action }}"
+        public-id="{{ optional($collection)->public_id }}"
+        update-token="{{ optional($collection)->update_token }}"
         :creating="@json($creating)"
         v-slot="form"
     >
@@ -46,31 +48,32 @@
             /></inline-input>
         </div>
     </div>
-    @if (session('created-message'))
-        <div class="row mt-5">
-            <div class="col-sm-6 col-sm-offset-3">
-                <div class="alert alert-info alert-dismissable mb-0">
-                    <button type="button" class="close" data-dismiss="alert" aria-hidden="true">&times;</button>
-                    {{ session('created-message') }}
-                </div>
-            </div>
-        </div>
-    @endif
     @if (!$disabled)
-        <div class="row mt-5" style="height:34px" v-cloak>
+        <div class="row mt-5" style="height:110px" v-cloak>
             <div class="col-sm-6 col-sm-offset-3 text-center">
                 <transition
                     enter-active-class="animated fadeInDown faster"
-                    leave-active-class="animated fadeOutUp faster"
+                    leave-active-class="animated fadeOut faster"
                     mode="out-in"
                 >
-                    <button v-if="form.editing" type="submit" class="btn btn-dark" key="save">
+                    <button v-if="form.editing" type="submit" class="btn btn-dark font-light p-3 px-5 mt-5" key="save">
                         uložiť úpravy {{-- TODO i18n --}}
                     </button>
                     @if (!$creating)
-                        <button v-if="!form.editing" type="button" class="btn btn-info" key="share" data-toggle="modal" data-target="#share">
-                            zdieľať výber <i class='ml-1 fa fa-share-alt'></i> {{-- TODO i18n --}}
-                        </button>
+                    <div v-else class="row bg-gray-300 py-5">
+                        <h5 class="text-xl font-semibold mt-0 mb-4">
+                            @if (session('user-collection-created'))Výborne! @endif
+                            Zdieľaj svoj výber s ostatnými:
+                        </h5>
+
+                        <div class="col-md-10 col-md-offset-1 mt-3">
+                            <copy-to-clipboard-group
+                                value="{{ route('frontend.shared-user-collections.show', $collection) }}"
+                                button-label="{{ trans('general.copy_link') }}"
+                                success-text="{{ trans('general.copied_to_clipboard') }}"
+                            ></copy-to-clipboard-group>
+                        </div>
+                    </div>
                     @endif
                 </transition>
             </div>
@@ -94,69 +97,6 @@
         @endforeach
     </div>
 </user-collections-share-form>
-
-@if (!$disabled && $collection)
-<div tabindex="-1" class="modal fade" id="share" role="dialog">
-    <div class="modal-dialog modal-md">
-        <div class="modal-content">
-            <div class="modal-body pb-5">
-                <div class="row">
-                    <div class="col-xs-12">
-                        <button type="button" class="close" data-dismiss="modal" aria-hidden="true">&times;</button>
-                    </div>
-                </div>
-                <div class="row">
-                    <div class="hidden-xs col-sm-2 text-right pr-0">
-                        <i class="fa fa-share-alt color mt-2 text-xl"></i>
-                    </div>
-                    <div class="col-sm-8">
-                        <h4>Odkaz na zdieľanie</h4>
-                        <p class="mt-3">
-                            Pošli tento odkaz svojim známym, alebo ho zdieľaj na sociálnych sieťach.
-                        </p>
-                        <copy-to-clipboard
-                            value="{{ route('frontend.shared-user-collections.show', $collection) }}"
-                            button-label="{{ trans('general.copy') }}"
-                            success-text="{{ trans('general.copied_to_clipboard') }}"
-                        ></copy-to-clipboard>
-
-
-                        <hr />
-                    </div>
-                </div>
-                <div class="row">
-                    <div class="hidden-xs col-sm-2 text-right pr-0">
-                        <i class="fa fa-pencil color mt-2 text-xl"></i>
-                    </div>
-                    <div class="col-sm-8">
-                        <h4>Odkaz na úpravy</h4>
-                        <copy-to-clipboard
-                            class="no-border mt-2"
-                            value="{{ route('frontend.shared-user-collections.edit', ['collection' => $collection, 'token' => $collection->update_token]) }}"
-                            button-label="{{ trans('general.copy') }}"
-                            success-text="{{ trans('general.copied_to_clipboard') }}"
-                        ></copy-to-clipboard>
-
-                        <div class="alert alert-info mt-3">
-                            Psst, tento odkaz si <strong>odlož</strong>. Môžeš ho použiť na ďalšie úpravy tohto výberu.
-                        </div>
-
-                        <hr />
-
-                        <p >
-                            Na vylepšovaní tejto funkcionality ešte stále pracujeme. Pomôž nám svojimi postrehmi a návrhmi
-                            na zlepšenie!
-                        </p>
-                        <p>
-                            Napíš nám na <a href="mailto:lab@sng.sk" class="underline">lab@sng.sk</a>
-                        </p>
-                    </div>
-                </div>
-            </div>
-        </div>
-    </div>
-</div>
-@endif
 
 @section('javascript')
 <script type="text/javascript">