+++ resources/views/frontend/user-collection/show.blade.php
@@ -9,27 +9,65 @@
 @section('content')
 <div class="container">
     <div class="row">
-        <div class="col-md-8 col-md-offset-2 col-lg-4 col-lg-offset-4">
-            <h2 class="bottom-space text-center">{{ trans('user-collection.title') }} <span class="badge badge-primary badge-sup">beta</span></h2>
+        <div class="col-md-6 col-md-offset-3">
+            <h2 class="bottom-space text-center font-semibold">
+                {{ trans('user-collection.title') }} <span class="badge badge-primary badge-sup">beta</span>
+            </h2>
 
-            @if($items->isEmpty())
-                <p class="alert alert-info text-center">{{ trans('user-collection.empty') }}</p>
-            @else
-                <p>{!! trans('user-collection.content-intro') !!}</p>
-                <p>{!! trans('user-collection.content-usage') !!}</p>
-                <p class="underlined-links">{!! trans('user-collection.content-feedback') !!}</p>
-
-                <div class="text-center">
+            <div class="text-lg mt-5">
+                @if($items->isNotEmpty())
+                    <p>{{ trans('user-collection.content-intro') }}</p>
+                    <p>{{ trans('user-collection.content-usage') }}</p>
                     @if (Experiment::is('WEBUMENIA-1654-shared-user-collections-with-scena'))
-                        <a class="btn btn-primary mt-5" href="{{ route('frontend.shared-user-collections.create', ['ids' => request()->ids ] ) }}">
-                            {{ trans('user-collection.share') }} <i class='ml-1 fa fa-share-alt'></i>
-                        </a>
+                        <p>{{ trans('user-collection.share-info') }}</p>
                     @endif
-                </div>
-            @endif
+                @else
+                    <p class="alert alert-info text-center">{{ trans('user-collection.empty') }}</p>
+                @endif
+            </div>
         </div>
     </div>
+</div>
 
+@if (Experiment::is('WEBUMENIA-1654-shared-user-collections-with-scena'))
+<user-collections-store v-slot="store">
+    <div :class="{ 'bg-gray-300': store.sharedCollections.present }" class="mt-5 py-5">
+        <div class="container">
+            <template v-if="store.sharedCollections.present">
+                <h2 class="font-semibold text-center m-0">Tvoje výbery</h2>
+
+                <div class="row mt-4">
+                    <div class="col-md-6 col-md-offset-3">
+                        <div class="bg-white p-4 mt-4" v-for="collection in store.sharedCollections.all" :key="collection.publicId">
+                            <user-collections-shared-collection
+                                    :public-id="collection.publicId"
+                                    :update-token="collection.updateToken"
+                                    api-url-template="{{ route('api.shared-user-collections.show', '__PUBLIC_ID__') }}"
+                                    share-url-template="{{ route('frontend.shared-user-collections.show', '__PUBLIC_ID__') }}"
+                                    edit-url-template="{{ route('frontend.shared-user-collections.edit', ['collection' => '__PUBLIC_ID__', 'token' => '__UPDATE_TOKEN__']) }}"
+                                    copy-success-text="{{ trans('general.copied_to_clipboard') }}"
+                                />
+                        </div>
+                    </div>
+                </div>
+            </template>
+
+            <div class="text-center mt-5">
+                <a class="btn btn-dark font-light p-3 px-5" href="{{ route('frontend.shared-user-collections.create', ['ids' => request()->ids ] ) }}">
+                    {{ trans('user-collection.share') }}
+                </a>
+
+                <div class="text-muted mt-5 mb-5">
+                    <p class="mb-0">Do výberu sa vložia všetky aktuálne obľúbené diela.</p>
+                    <p v-if="store.sharedCollections.present" class="mt-1 mb-0">Diela v ostatných výberoch ostanú zachované.</p>
+                </div>
+            </div>
+        </div>
+    </div>
+</user-collections-store>
+@endif
+
+<div class="container">
     @unless($items->isEmpty())
     <div class="row content-section">
         <div class="col-xs-6">